requirements.txt
pyserial
argparse
numpy